OrderedDict, BTree and dict all iterate the same way.
Fixes #23. Also a further fix for #20 (you couldn't iterate a BTree
all by itself).

Refactor the test case for BTree to be a shared implementation and
confirm that it works as expected for dict, using the actual dict
checker. Then apply it to OrderedDict and BTree and fix the resulting
failures by refactoring the fixup in checker.py to a shared
implementation and applying it.