Switch to modern 'coverage' environment and run doctests on all versions
Based on #29.

Drop Python 3.3.

Enable coveralls on Travis CI, and also fix using interpreters that
aren't there in the new default 'trusty' image.

Coverage is not at 100% and in fact appears to have dropped (I show
92% in one run, vs the previous reported 97%), but that's because
`coverage` and coveralls report files that aren't even imported (we
have a lot of those in the 'tests' package)...nosetests didn't.