Fix default values for Permission's title and description in Py2
And test this. This applies to both the class and the ZCML directive.

Also adjust IPermission to use NativeStringLine for ``id``, since
that's what ZCML has always been creating (it uses an Id, which
subclasses NativeStringLine), and what's typically written in source
code. Test that the permissions created from ZCML are valid with this
change.