Remove obsolete tests.
I renamed the tests (that's why their name ends with `2`) there where methods
with the same names around line 1600.

The other tests assure that a TypeError is raised when calling
`pow(x, y, proxy)` because "a PyProxy cannot lie about its type".

Interestingly the deleted tests did not fail on CPython but the tests expecting
the TypeError were locally successful, too, but not at GHA – that's strange.

I am restoring here the status before my renaming of the tests: The now deleted
tests did not run because of the name clash.